+++ java-client/src/test/java/org/opensearch/client/opensearch/experiments/ParsingTests.java
@@ -33,13 +33,15 @@
 package org.opensearch.client.opensearch.experiments;
 
 import jakarta.json.spi.JsonProvider;
-import jakarta.json.stream.JsonGenerator;
-import jakarta.json.stream.JsonParser;
 import jakarta.json.stream.JsonParsingException;
 import org.junit.Assert;
 import org.junit.Test;
+import org.opensearch.client.json.JsonpDeserializer;
+import org.opensearch.client.json.JsonpSerializable;
 import org.opensearch.client.json.jsonb.JsonbJsonpMapper;
+import org.opensearch.client.opensearch._types.Time;
 import org.opensearch.client.opensearch.experiments.api.FooRequest;
+import org.opensearch.client.opensearch.indices.IndexSettings;
 import org.opensearch.client.opensearch.indices.IndexSettingsMapping;
 
 import java.io.ByteArrayInputStream;
@@ -53,8 +55,6 @@ public class ParsingTests extends Assert {
 
     try {
 
-      ByteArrayOutputStream baos = new ByteArrayOutputStream();
-
       FooRequest foo = FooRequest.builder()
               .name("z")
               .value(1)
@@ -64,19 +64,10 @@ public class ParsingTests extends Assert {
               )
               .build();
 
-      JsonProvider provider = JsonProvider.provider();
-      JsonGenerator generator = provider.createGenerator(baos);
-      foo.serialize(generator, new JsonbJsonpMapper());
-
-      generator.close();
-      String str = baos.toString();
-
+      String str = serialize(foo);
       assertEquals("{\"name\":\"z\",\"value\":1,\"indices\":[\"a\",\"b\",\"c\"],\"bar\":{\"name\":\"Raise the bar\"}}", str);
 
-      JsonParser parser = provider.createParser(new ByteArrayInputStream(str.getBytes(StandardCharsets.UTF_8)));
-
-      FooRequest foo2 = FooRequest.parser().deserialize(parser, new JsonbJsonpMapper());
-
+      FooRequest foo2 = deserialize(str, FooRequest.parser());
       assertEquals(foo.name(), foo2.name());
       assertEquals(foo.value(), foo2.value());
       assertNull(foo2.size());
@@ -88,11 +79,31 @@ public class ParsingTests extends Assert {
   }
 
   @Test
-  public void testIndexSettingsMappingParsing() {
+  public void testIndexSettingsSyncIntervalTimeParsing() {
 
     try {
 
-      var baos = new ByteArrayOutputStream();
+      var indexSettings = IndexSettings.of(_1 -> _1.translogSyncInterval(Time.of(_2 -> _2.time("10s"))));
+
+      var str = serialize(indexSettings);
+      assertEquals("{\"translog.sync_interval\":\"10s\"}", str);
+
+      IndexSettings deserialized = deserialize(str, IndexSettings._DESERIALIZER);
+      assertEquals(indexSettings.translogSyncInterval().time(), deserialized.translogSyncInterval().time());
+
+      var responseStr = "{\"translog\":{\"sync_interval\":\"10s\"}}";
+      IndexSettings responseDeserialized = deserialize(responseStr, IndexSettings._DESERIALIZER);
+      assertEquals(indexSettings.translogSyncInterval().time(), responseDeserialized.translog().syncInterval().time());
+
+    } catch (JsonParsingException je) {
+      throw new JsonParsingException(je.getMessage() + " at " + je.getLocation(), je, je.getLocation());
+    }
+  }
+
+  @Test
+  public void testIndexSettingsMappingParsing() {
+
+    try {
 
       var mapping = IndexSettingsMapping.of(b -> b
               .totalFields(d -> d.limit(1L))
@@ -101,20 +112,11 @@ public class ParsingTests extends Assert {
               .nestedObjects(d -> d.limit(4L))
               .fieldNameLength(d -> d.limit(5L)));
 
-      var provider = JsonProvider.provider();
-      var generator = provider.createGenerator(baos);
-      mapping.serialize(generator, new JsonbJsonpMapper());
-
-      generator.close();
-      var str = baos.toString();
-
+      var str = serialize(mapping);
       assertEquals("{\"total_fields\":{\"limit\":1},\"depth\":{\"limit\":2},\"nested_fields\":{\"limit\":3}," +
                    "\"nested_objects\":{\"limit\":4},\"field_name_length\":{\"limit\":5}}", str);
 
-      var parser = provider.createParser(new ByteArrayInputStream(str.getBytes(StandardCharsets.UTF_8)));
-
-      var deserialized = IndexSettingsMapping._DESERIALIZER.deserialize(parser, new JsonbJsonpMapper());
-
+      var deserialized = deserialize(str, IndexSettingsMapping._DESERIALIZER);
       assertEquals(mapping.totalFields().limit(), deserialized.totalFields().limit());
       assertEquals(mapping.depth().limit(), deserialized.depth().limit());
       assertEquals(mapping.nestedFields().limit(), deserialized.nestedFields().limit());
@@ -123,5 +125,23 @@ public class ParsingTests extends Assert {
     } catch (JsonParsingException je) {
       throw new JsonParsingException(je.getMessage() + " at " + je.getLocation(), je, je.getLocation());
     }
+  }
+
+  private <T extends JsonpSerializable> T deserialize(String serializedValue, JsonpDeserializer<T> deserializer) {
+    var provider = JsonProvider.provider();
+    var parser = provider.createParser(new ByteArrayInputStream(serializedValue.getBytes(StandardCharsets.UTF_8)));
+
+    return deserializer.deserialize(parser, new JsonbJsonpMapper());
+  }
+
+  private String serialize(JsonpSerializable object) {
+    var baos = new ByteArrayOutputStream();
+    var provider = JsonProvider.provider();
+
+    var generator = provider.createGenerator(baos);
+    object.serialize(generator, new JsonbJsonpMapper());
+    generator.close();
+
+    return baos.toString();
   }
 }